pyproject.toml
[project]
name = "awesome-fl-seg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cityscapesscripts>=2.2.4",
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "numpy==1.22.4",
    "opencv-python",
    "pandas>=1.5.1",
    "segmentation-models-pytorch>=0.5.0",
    "torch==1.12.1+cu113",
    "torchaudio==0.12.1",
    "torchvision==0.13.1+cu113",
    "transformers==4.33.2",
    "wandb>=0.21.1",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu113"

[[tool.uv.index]]
url = "https://pypi.org/simple"

